+++ core/generator.go
