+++ core/values.go
